Made Joboffers a module
