+++ amivapi/bootstrap.py
@@ -33,7 +33,8 @@ from amivapi import (
     groups,
     ldap,
     documentation,
-    utils
+    utils,
+    joboffers
 )
 
 from amivapi.utils import get_config
@@ -93,6 +94,7 @@ def create_app(disable_auth=False, **kwargs):
     auth.init_app(app)
     events.init_app(app)
     groups.init_app(app)
+    joboffers.init_app(app)
     media.init_app(app)
 
     return app
